Additional refactoring of code from RemoteControllerWebSocket to WebSocket handlers
